requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.11.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
qrcode==8.2
reportlab==4.4.3
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
